+++ tests/konsist/src/test/kotlin/io/element/android/tests/konsist/KonsistArchitectureTest.kt
@@ -19,6 +19,7 @@ import com.lemonappdev.konsist.api.ext.list.withNameEndingWith
 import com.lemonappdev.konsist.api.ext.list.withoutAnnotationOf
 import com.lemonappdev.konsist.api.ext.list.withoutConstructors
 import com.lemonappdev.konsist.api.ext.list.withoutName
+import com.lemonappdev.konsist.api.ext.list.withoutPackage
 import com.lemonappdev.konsist.api.ext.list.withoutParents
 import com.lemonappdev.konsist.api.verify.assertEmpty
 import com.lemonappdev.konsist.api.verify.assertTrue
@@ -35,6 +36,9 @@ class KonsistArchitectureTest {
             .withoutName(
                 "CameraPositionState",
                 "CustomSheetState",
+            )
+            .withoutPackage(
+                "de.bwi.messenger.features.messages.impl.timeline.components.event"
             )
             .constructors
             .parameters
