+++ tests/konsist/src/test/kotlin/io/element/android/tests/konsist/KonsistCallbackTest.kt
@@ -8,6 +8,7 @@
 package io.element.android.tests.konsist
 
 import com.lemonappdev.konsist.api.Konsist
+import com.lemonappdev.konsist.api.ext.list.withoutPackage
 import com.lemonappdev.konsist.api.verify.assertFalse
 import org.junit.Test
 
@@ -17,6 +18,9 @@ class KonsistCallbackTest {
         Konsist
             .scopeFromProduction()
             .files
+            .withoutPackage(
+                "de.bwi.messenger.features.messages.impl.timeline.components.event"
+            )
             .assertFalse {
                 it.text.contains("callback?.")
             }
