+++ tests/konsist/src/test/kotlin/io/element/android/tests/konsist/KonsistClassNameTest.kt
@@ -18,6 +18,7 @@ import com.lemonappdev.konsist.api.ext.list.withNameEndingWith
 import com.lemonappdev.konsist.api.ext.list.withPackage
 import com.lemonappdev.konsist.api.ext.list.withoutName
 import com.lemonappdev.konsist.api.ext.list.withoutNameStartingWith
+import com.lemonappdev.konsist.api.ext.list.withoutPackage
 import com.lemonappdev.konsist.api.verify.assertEmpty
 import com.lemonappdev.konsist.api.verify.assertTrue
 import io.element.android.libraries.architecture.Presenter
@@ -29,6 +30,9 @@ class KonsistClassNameTest {
         Konsist.scopeFromProject()
             .classes()
             .withAllParentsOf(Presenter::class)
+            .withoutPackage(
+                "de.bwi.messenger.features.messages.impl.timeline.components.event"
+            )
             .assertTrue {
                 it.name.endsWith("Presenter")
             }
@@ -39,6 +43,9 @@ class KonsistClassNameTest {
         Konsist.scopeFromProject()
             .classes()
             .withAllParentsOf(Node::class)
+            .withoutPackage(
+                "de.bwi.messenger.features.messages.impl.timeline.components.event"
+            )
             .assertTrue {
                 it.name.endsWith("Node")
             }
@@ -53,6 +60,9 @@ class KonsistClassNameTest {
                 "AspectRatioProvider",
                 "OverlapRatioProvider",
                 "TextFileContentProvider",
+            )
+            .withoutPackage(
+                "de.bwi.messenger.features.messages.impl.timeline.components.event"
             )
             .also {
                 // Check that classes are actually found
@@ -87,6 +97,9 @@ class KonsistClassNameTest {
             .withoutName(
                 "FakeFileSystem",
                 "FakeImageLoader",
+            )
+            .withoutPackage(
+                "de.bwi.messenger.features.messages.impl.timeline.components.event"
             )
             .assertTrue {
                 val interfaceName = it.name
@@ -124,6 +137,9 @@ class KonsistClassNameTest {
             .classes()
             .withNameEndingWith("Impl")
             .withoutName("MediaUploadHandlerImpl")
+            .withoutPackage(
+                "de.bwi.messenger.features.messages.impl.timeline.components.event"
+            )
             .assertEmpty(additionalMessage = "Class implementing interface should have name not end with 'Impl' but start with 'Default'")
     }
 
@@ -157,6 +173,9 @@ class KonsistClassNameTest {
                 "Preferences",
                 "Rust",
                 "SharedPreferences",
+            )
+            .withoutPackage(
+                "de.bwi.messenger.features.messages.impl.timeline.components.event"
             )
             .assertEmpty()
     }
