+++ tests/konsist/src/test/kotlin/io/element/android/tests/konsist/KonsistFieldTest.kt
@@ -9,6 +9,7 @@ package io.element.android.tests.konsist
 
 import com.lemonappdev.konsist.api.Konsist
 import com.lemonappdev.konsist.api.ext.list.properties
+import com.lemonappdev.konsist.api.ext.list.withoutPackage
 import com.lemonappdev.konsist.api.verify.assertFalse
 import org.junit.Test
 
@@ -18,6 +19,9 @@ class KonsistFieldTest {
         Konsist
             .scopeFromProject()
             .classes()
+            .withoutPackage(
+                "de.bwi.messenger.features.messages.impl.timeline.components.event"
+            )
             .properties()
             .assertFalse {
                 val secondCharacterIsUppercase = it.name.getOrNull(1)?.isUpperCase() ?: false
