+++ tests/konsist/src/test/kotlin/io/element/android/tests/konsist/KonsistMethodNameTest.kt
@@ -8,6 +8,7 @@
 package io.element.android.tests.konsist
 
 import com.lemonappdev.konsist.api.Konsist
+import com.lemonappdev.konsist.api.ext.list.withoutPackage
 import com.lemonappdev.konsist.api.verify.assertTrue
 import org.junit.Test
 
@@ -16,6 +17,9 @@ class KonsistMethodNameTest {
     fun `Ensure that method name does not start or end with spaces`() {
         Konsist.scopeFromProject()
             .functions()
+            .withoutPackage(
+                "de.bwi.messenger.features.messages.impl.timeline.components.event"
+            )
             .assertTrue {
                 it.name.trim() == it.name
             }
