+++ tests/konsist/src/test/kotlin/io/element/android/tests/konsist/KonsistParameterNameTest.kt
@@ -9,6 +9,7 @@ package io.element.android.tests.konsist
 
 import com.lemonappdev.konsist.api.Konsist
 import com.lemonappdev.konsist.api.ext.list.withParameter
+import com.lemonappdev.konsist.api.ext.list.withoutPackage
 import com.lemonappdev.konsist.api.verify.assertEmpty
 import org.junit.Test
 
@@ -20,6 +21,9 @@ class KonsistParameterNameTest {
             .withParameter { parameter ->
                 parameter.name.endsWith("Press")
             }
+            .withoutPackage(
+                "de.bwi.messenger.features.messages.impl.timeline.components.event"
+            )
             .assertEmpty(additionalMessage = "Please rename the parameter, for instance from 'onBackPress' to 'onBackClick'.")
     }
 }
